requirements.txt
asgiref==3.5.2
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.6
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.13.1
drf-api-logger==1.1.10
drf-yasg2==1.19.4
executing==0.8.3
graphviz==0.20
idna==3.3
inflection==0.5.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jwcrypto==1.3.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysqlclient==2.1.1
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
